Refactor events service for improved type safety and maintainability
- Add explicit PostgreSQL type casting to all SQL queries for better type safety
- Simplify events service by removing concurrent processing and circuit breakers
- Update SQLC generated code with concrete Go types instead of interfaces
- Remove sony/gobreaker dependency to reduce complexity
- Enable MMA and Tennis for detailed odds fetching (previously skipped)
- Update documentation with current API endpoints and production guidance
- Fix pagination type conversion in events handler

This refactoring prioritizes code simplicity and maintainability while
maintaining core functionality. The sequential processing approach makes
debugging easier, though may impact performance for bulk operations.

+++ pkg/handlers/events/handler.go
@@ -151,7 +151,7 @@ func (h *Handler) List(w http.ResponseWriter, r *http.Request) {
 	events := h.convertEventsToResponse(dbEvents)
 
 	// Calculate pagination metadata
-	totalPages := int((totalCount + int64(perPage) - 1) / int64(perPage)) // Ceiling division
+	totalPages := int((int64(totalCount) + int64(perPage) - 1) / int64(perPage)) // Ceiling division
 	hasNext := page < totalPages
 	hasPrevious := page > 1
 
@@ -179,7 +179,7 @@ func (h *Handler) List(w http.ResponseWriter, r *http.Request) {
 		Str("status", status).
 		Int("page", page).
 		Int("per_page", perPage).
-		Int64("total", totalCount).
+		Int("total", int(totalCount)).
 		Int("count", len(events)).
 		Msg("Returning paginated events")
